+++ src/main/java/de/codescape/jira/plugins/scrumpoker/ao/ScrumPokerSession.java
@@ -83,4 +83,11 @@ public interface ScrumPokerSession extends RawEntity<String> {
 
     void setCancelled(boolean cancelled);
 
+    /**
+     * The card set used by this Scrum Poker session.
+     */
+    String getCardSet();
+
+    void setCardSet(String cardSet);
+
 }