+++ src/main/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerCardServiceImpl.java
@@ -1,9 +1,6 @@
 package de.codescape.jira.plugins.scrumpoker.service;
 
-import com.atlassian.activeobjects.external.ActiveObjects;
-import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
-import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerCards;
-import net.java.ao.Query;
+import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -17,20 +14,21 @@ import java.util.stream.Collectors;
 @Component
 public class ScrumPokerCardServiceImpl implements ScrumPokerCardService {
 
-    private final ActiveObjects activeObjects;
+    private final ScrumPokerSettingService scrumPokerSettingService;
 
     @Autowired
-    public ScrumPokerCardServiceImpl(@ComponentImport ActiveObjects activeObjects) {
-        this.activeObjects = activeObjects;
+    public ScrumPokerCardServiceImpl(ScrumPokerSettingService scrumPokerSettingService) {
+        this.scrumPokerSettingService = scrumPokerSettingService;
     }
 
     @Override
     public List<String> getCardSet() {
-        // at the moment we only expect one entry but to make sure we only query for one
-        ScrumPokerCards[] cardSets = activeObjects.find(ScrumPokerCards.class, Query.select()
-            .order("ID DESC")
-            .limit(1));
-        return splitToCards(cardSets[0].getCardSet());
+        return splitToCards(scrumPokerSettingService.load().getCardSet());
+    }
+
+    @Override
+    public List<String> getCardSet(ScrumPokerSession scrumPokerSession) {
+        return scrumPokerSession.getCardSet() != null ? splitToCards(scrumPokerSession.getCardSet()) : getCardSet();
     }
 
     private List<String> splitToCards(String cardSet) {