+++ src/test/java/de/codescape/jira/plugins/scrumpoker/ScrumPokerTestDatabaseUpdater.java
@@ -13,7 +13,6 @@ public class ScrumPokerTestDatabaseUpdater implements DatabaseUpdater {
     public void update(EntityManager entityManager) throws Exception {
         //noinspection unchecked
         entityManager.migrate(
-            ScrumPokerCards.class,
             ScrumPokerError.class,
             ScrumPokerProject.class,
             ScrumPokerSession.class,