+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerCardServiceImplTest.java
@@ -1,8 +1,6 @@
 package de.codescape.jira.plugins.scrumpoker.service;
 
-import com.atlassian.activeobjects.external.ActiveObjects;
-import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerCards;
-import net.java.ao.Query;
+import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
 import org.hamcrest.Matchers;
 import org.junit.Rule;
 import org.junit.Test;
@@ -15,7 +13,6 @@ import java.util.List;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 public class ScrumPokerCardServiceImplTest {
@@ -24,13 +21,13 @@ public class ScrumPokerCardServiceImplTest {
     public MockitoRule rule = MockitoJUnit.rule();
 
     @Mock
-    private ActiveObjects activeObjects;
+    private ScrumPokerSettingService scrumPokerSettingService;
 
     @InjectMocks
     private ScrumPokerCardServiceImpl service;
 
     @Mock
-    private ScrumPokerCards scrumPokerCards;
+    private GlobalSettings globalSettings;
 
     @Test
     public void shouldIgnoreWhitespaceOnElementsInCardSet() {
@@ -49,10 +46,8 @@ public class ScrumPokerCardServiceImplTest {
     }
 
     private void expectResultFromActiveObjects(String expectedString) {
-        when(scrumPokerCards.getCardSet()).thenReturn(expectedString);
-        ScrumPokerCards[] scrumPokerCards = {this.scrumPokerCards};
-        // noinspection unchecked
-        when(activeObjects.find(any(Class.class), any(Query.class))).thenReturn(scrumPokerCards);
+        when(scrumPokerSettingService.load()).thenReturn(globalSettings);
+        when(globalSettings.getCardSet()).thenReturn(expectedString);
     }
 
 }